Refactored configuration code into AndroidModulePlugin
